Move index permissions tests to packages
