--- index_permissions_test.go
@@ -1,66 +1,0 @@
-package goa4web
-
-import (
-	"net/http/httptest"
-	"testing"
-
-	corecommon "github.com/arran4/goa4web/core/common"
-	blogs "github.com/arran4/goa4web/handlers/blogs"
-	news "github.com/arran4/goa4web/handlers/news"
-)
-
-func TestCustomNewsIndexRoles(t *testing.T) {
-	req := httptest.NewRequest("GET", "/", nil)
-
-	cd := &CoreData{SecurityLevel: "administrator"}
-	news.CustomNewsIndex(cd, req)
-	if !corecommon.ContainsItem(cd.CustomIndexItems, "User Permissions") {
-		t.Errorf("admin should see user permissions")
-	}
-	if !corecommon.ContainsItem(cd.CustomIndexItems, "Add News") {
-		t.Errorf("admin should see add news")
-	}
-
-	cd = &CoreData{SecurityLevel: "writer"}
-	news.CustomNewsIndex(cd, req)
-	if corecommon.ContainsItem(cd.CustomIndexItems, "User Permissions") {
-		t.Errorf("writer should not see user permissions")
-	}
-	if !corecommon.ContainsItem(cd.CustomIndexItems, "Add News") {
-		t.Errorf("writer should see add news")
-	}
-
-	cd = &CoreData{SecurityLevel: "reader"}
-	news.CustomNewsIndex(cd, req)
-	if corecommon.ContainsItem(cd.CustomIndexItems, "User Permissions") || corecommon.ContainsItem(cd.CustomIndexItems, "Add News") {
-		t.Errorf("reader should not see admin items")
-	}
-}
-
-func TestCustomBlogIndexRoles(t *testing.T) {
-	req := httptest.NewRequest("GET", "/blogs", nil)
-
-	cd := &CoreData{SecurityLevel: "administrator"}
-	blogs.CustomBlogIndex(cd, req)
-	if !corecommon.ContainsItem(cd.CustomIndexItems, "User Permissions") {
-		t.Errorf("admin should see user permissions")
-	}
-	if !corecommon.ContainsItem(cd.CustomIndexItems, "Write blog") {
-		t.Errorf("admin should see write blog")
-	}
-
-	cd = &CoreData{SecurityLevel: "writer"}
-	blogs.CustomBlogIndex(cd, req)
-	if corecommon.ContainsItem(cd.CustomIndexItems, "User Permissions") {
-		t.Errorf("writer should not see user permissions")
-	}
-	if !corecommon.ContainsItem(cd.CustomIndexItems, "Write blog") {
-		t.Errorf("writer should see write blog")
-	}
-
-	cd = &CoreData{SecurityLevel: "reader"}
-	blogs.CustomBlogIndex(cd, req)
-	if corecommon.ContainsItem(cd.CustomIndexItems, "User Permissions") || corecommon.ContainsItem(cd.CustomIndexItems, "Write blog") {
-		t.Errorf("reader should not see writer/admin items")
-	}
-}
